Use ptrdiff_t for safe pointer subtraction and size casting
Fix type correctness in `android-blob-utility.c` by switching to `ptrdiff_t` and casting lengths to avoid signed/size truncation when computing substrings for `strncpy`. Update the Windows mmap compatibility shim to expand the POSIX/GNU API coverage (add `getline`, define `ssize_t`) and improve documentation, ensuring the utility builds cleanly on MinGW.

+++ android-blob-utility.c
@@ -412,7 +412,7 @@ void get_full_lib_name(char *found_lib) {
 
     char full_name[512] = {0};
 
-    long len;
+    ptrdiff_t len;
     int num_chars;
     int i;
 
@@ -473,8 +473,8 @@ void get_full_lib_name(char *found_lib) {
         ptr--;
         peek--;
     }
-    len = (long)(found_lib + strlen(lib_beginning)) - (long)ptr;
-    strncpy(full_name, ptr, len);
+    len = (found_lib + (ptrdiff_t)strlen(lib_beginning)) - ptr;
+    strncpy(full_name, ptr, (size_t)len);
 
     check_emulator_for_lib(full_name);
 }
